requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
python-dateutil>=2.8.0